Data length correction / memory leaks fixed

+++ Source/Characters/LivingEntity.cpp
@@ -26,9 +26,9 @@ void LivingEntity::drawHealthBar(RenderContext& renderer, int x, int y)
 	int X = x - sz4;
 	int Y = y - sz2;
 	renderer.changeColor(0xFFFFFFFF);
-	renderer.drawRectangle(X, Y, SZ_BLOCKSIZE * 1.5 - 1, sz4 - 1, true);
+	renderer.drawRectangle(X, Y, int(SZ_BLOCKSIZE * 1.5) - 1, sz4 - 1, true);
 	renderer.changeColor(0x00FF00FF);
-	renderer.drawRectangle(X, Y, health * (SZ_BLOCKSIZE * 1.5) / maxHealth - 1, sz4, true);
+	renderer.drawRectangle(X, Y, int(health * SZ_BLOCKSIZE * 1.5 / maxHealth) - 1, sz4, true);
 	renderer.changeColor(0x00000000);
-	renderer.drawRectangle(X - 1, Y - 1, SZ_BLOCKSIZE * 1.5 + 1, sz4 + 1);
+	renderer.drawRectangle(X - 1, Y - 1, int(SZ_BLOCKSIZE * 1.5) + 1, sz4 + 1);
 }
